BACKEND/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
azure-ai-vision==0.15.1b1
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.28
azure-core==1.29.6
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.16.1
dill==0.3.7
etelemetry==0.3.1
faiss-cpu==1.7.4
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-vision==3.5.0
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
huggingface-hub==0.20.2
idna==3.6
imutils==0.5.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.11
langchain-google-genai==0.0.6
langsmith==0.0.81
looseversion==1.3.0
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
pdf2image==1.16.3
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.2
prov==2.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traits==6.3.2
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
